# openapi.yaml
openapi: 3.0.1
info:
  title: API da Atividade 3
  description: API para a Atividade 3.
  version: 0.0.1
  termsOfService: https://mockapi.io
  contact:
    name: Suporte a Devs
    email: contato@example.com
    url: https://mockapi.io
  license:
    name: 'Lincença: GPLv3'
    url: https://www.gnu.org/licenses/gpl-3.0.html
externalDocs:
  description: Documentação burocrática
  url: https://mockapi.io
servers:
  - url: https://localhost:8080/estados
    description: Atividade 3
paths:
  /estados:
    get:
      summary: traz todos os estados criados
      parameters:
        - in: path
          name: Regiao
          schema:
            type: string
          required: false
          description: Pega os estados por Regiao
        - in: path
          name: Populacao
          schema:
            type: integer
          required: false
          description: Organiza os estados por Populacao
        - in: path
          name: Area
          schema:
            type: integer
          required: false
          description: Organiza os estados por Area
      responses:
        '200':
          description: Sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Estados'
    post:
      summary: Cria novo estado
      requestBody:
        description: Optional description in *Markdown*
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Estados'
      responses:
        '201':
          description: Sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Estados'
  /estados/{id}:
    parameters:
      - name: id
        in: path
        schema:
          type: integer
        required: true
    get:
      summary: Recupera um estado pelo ID
      responses:
        '200':
          description: Sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Estados'
        '404':
          description: Estado não encontrada
          content:
            application/json:
              example: Not Found
    put:
      summary: Atualiza um estado
      description: Atualiza um estado pelo ID
      operationId: id
      responses:
        default:
          description: Sucesso
          
    delete:
      summary: Deleta um estado
      description: Atualiza um estado pelo ID
      operationId: ''
      responses:
        default:
          description: Sucesso
security:
  - auth: []
components:
  schemas:
    Estados:
      type: object
      properties:
        id:
          type: integer
        Nome:
          type: string
        Regiao:
          type: string
        Populacao:
          type: integer
        area:
          type: integer
        dataFundacao:
          type: string
        dataDesdeAfundacao:
          type: string
  securitySchemes:
    auth:
      type: http
      scheme: bearer
      bearerFormat: JWT